+++ tests/settings-docker.py
@@ -1,5 +1,4 @@
 #!/usr/bin/env python
-# -*- coding: utf-8 -*-
 import os
 import sys
 
@@ -45,7 +44,7 @@ HELPER_SETTINGS = {
         ('zh-cn', 'Chinese (Simplified)'),
     ),
     'LANGUAGE_CODE': 'en',
-    'ALLOWED_HOSTS': ['localhost'],
+    'ALLOWED_HOSTS': ['*'],
     'USE_TZ': True,
     'TIME_ZONE': 'Europe/Zurich',
     'PARLER_LANGUAGES': {