+++ CHANGELOG.rst
@@ -3,6 +3,16 @@ Changelog
 =========
 
 
+2.0.0 (unreleased)
+==================
+
+* Added support for Django 3.1
+* Dropped support for Python 2.7 and Python 3.4
+* Dropped support for Django < 2.2
+* Changed ``field-box`` class declarations to ``fieldBox``
+* Fixed screenshot tests for Django 2.2 and higher
+
+
 1.5.0 (2020-01-24)
 ==================